Changed content_length to be a signed integer - so it is possible to distinguish between a body with an unknown-length and no-length. Added redhttp_response_new_empty() function to create a response with zero length.

+++ src/redhttp/redhttp_private.h
@@ -79,7 +79,7 @@ struct redhttp_response_s {
   unsigned int status_code;
   char *status_message;
   char *content_buffer;
-  size_t content_length;
+  int content_length;
   void (*content_free_callback) (void *ptr);
 
   void *user_data;
